ALLL

+++ OrderService/Models/CustomerOrder.cs
@@ -7,7 +7,7 @@ public record CustomerOrder {
     public int Quantity  { get; set; }
     public double Amount { get; set; }
     public string PaymentMethod  { get; set; } = null!;
-    public long OrderId  { get; set; }
+    public int OrderId  { get; set; }
     public string? Address  { get; set; }
     public override string ToString() => JsonSerializer.Serialize(this);
 }
